fix(x11): Use int cast for IME spot location coordinates
The internal UpdateSpotLocationFromTextBox caller was still casting to
short before passing to UpdateSpotLocation(int, int). The clamping to
short range happens inside UpdateSpotLocation for the XIM path only.

+++ src/Uno.UI.Runtime.Skia.X11/UI/Xaml/Controls/TextBox/X11ImeTextBoxExtension.cs
@@ -252,7 +252,7 @@ internal sealed class X11ImeTextBoxExtension : IImeTextBoxExtension
 		var point = transform.TransformPoint(new Windows.Foundation.Point(rect.Left, rect.Top + rect.Height));
 		var scale = textBox.XamlRoot.RasterizationScale;
 
-		UpdateSpotLocation((short)(point.X * scale), (short)(point.Y * scale));
+		UpdateSpotLocation((int)(point.X * scale), (int)(point.Y * scale));
 	}
 
 	/// <summary>
